@@ -2,10 +2,10 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net5</TargetFramework>
         <RootNamespace>FastEnumUtility.Benchmark</RootNamespace>
-        <LangVersion>8.0</LangVersion>
-        <Nullable>disable</Nullable>
+        <LangVersion>9.0</LangVersion>
+        <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
