@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Runtime.CompilerServices;
 using System.Runtime.Serialization;
 
 
@@ -17,6 +18,7 @@ namespace FastEnumUtility
         /// <typeparam name="T"></typeparam>
         /// <param name="value"></param>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static Member<T> ToMember<T>(this T value)
             where T : struct, Enum
             => FastEnum.GetMember(value);
@@ -28,6 +30,7 @@ namespace FastEnumUtility
         /// <typeparam name="T"></typeparam>
         /// <param name="value"></param>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static string ToName<T>(this T value)
             where T : struct, Enum
             => FastEnum.GetName(value);
@@ -39,6 +42,7 @@ namespace FastEnumUtility
         /// <typeparam name="T"></typeparam>
         /// <param name="value"></param>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool IsDefined<T>(this T value)
             where T : struct, Enum
             => FastEnum.IsDefined(value);
@@ -51,16 +55,17 @@ namespace FastEnumUtility
         /// <param name="value"></param>
         /// <param name="throwIfNotFound"></param>
         /// <returns></returns>
-        public static string GetEnumMemberValue<T>(this T value, bool throwIfNotFound = true)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static string? GetEnumMemberValue<T>(this T value, bool throwIfNotFound = true)
             where T : struct, Enum
         {
             var attr = value.ToMember().EnumMemberAttribute;
-            if (attr != null)
+            if (attr is not null)
                 return attr.Value;
 
             return throwIfNotFound
                 ? throw new NotFoundException($"{nameof(EnumMemberAttribute)} is not found.")
-                : default(string);
+                : default;
         }
 
 
@@ -71,10 +76,11 @@ namespace FastEnumUtility
         /// <param name="member"></param>
         /// <param name="throwIfNotFound"></param>
         /// <returns></returns>
-        public static string GetLabel<T>(this Member<T> member, int index = 0, bool throwIfNotFound = true)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static string? GetLabel<T>(this Member<T> member, int index = 0, bool throwIfNotFound = true)
             where T : struct, Enum
         {
-            if (member == null)
+            if (member is null)
                 throw new ArgumentNullException(nameof(member));
 
             if (member.Labels.TryGetValue(index, out var label))
@@ -82,7 +88,7 @@ namespace FastEnumUtility
 
             return throwIfNotFound
                 ? throw new NotFoundException($"{nameof(LabelAttribute)} that is specified index {index} is not found.")
-                : default(string);
+                : default;
         }
 
 
@@ -93,7 +99,8 @@ namespace FastEnumUtility
         /// <param name="value"></param>
         /// <param name="throwIfNotFound"></param>
         /// <returns></returns>
-        public static string GetLabel<T>(this T value, int index = 0, bool throwIfNotFound = true)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static string? GetLabel<T>(this T value, int index = 0, bool throwIfNotFound = true)
             where T : struct, Enum
             => value.ToMember().GetLabel(index, throwIfNotFound);
     }
