@@ -98,7 +98,7 @@ namespace FastEnumUtility.Internals
                 => this.source[this.index];
 
 
-            object IEnumerator.Current
+            object? IEnumerator.Current
                 => this.Current;
 
 
@@ -144,7 +144,7 @@ namespace FastEnumUtility.Internals
                 => this.source[this.index];
 
 
-            object IEnumerator.Current
+            object? IEnumerator.Current
                 => this.Current;
 
 
@@ -181,12 +181,12 @@ namespace FastEnumUtility.Internals
         /// <returns></returns>
         public static ReadOnlyArray<T> ToReadOnlyArray<T>(this IEnumerable<T> source)
         {
-            if (source == null)
+            if (source is null)
                 throw new ArgumentNullException(nameof(source));
 
             return source is T[] array
-                ? new ReadOnlyArray<T>(array)
-                : new ReadOnlyArray<T>(source.ToArray());
+                ? new(array)
+                : new(source.ToArray());
         }
 
 
@@ -197,6 +197,6 @@ namespace FastEnumUtility.Internals
         /// <param name="source"></param>
         /// <returns></returns>
         public static ReadOnlyArray<T> AsReadOnly<T>(this T[] source)
-            => new ReadOnlyArray<T>(source);
+            => new(source);
     }
 }
