@@ -34,6 +34,7 @@ namespace FastEnumUtility
         /// </summary>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static IReadOnlyList<T> GetValues<T>()
             where T : struct, Enum
             => Cache_Values<T>.Values;
@@ -46,6 +47,7 @@ namespace FastEnumUtility
         /// </summary>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static IReadOnlyList<string> GetNames<T>()
             where T : struct, Enum
             => Cache_Names<T>.Names;
@@ -69,6 +71,7 @@ namespace FastEnumUtility
         /// </summary>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static IReadOnlyList<Member<T>> GetMembers<T>()
             where T : struct, Enum
             => Cache_Members<T>.Members;
@@ -116,6 +119,7 @@ namespace FastEnumUtility
         /// </summary>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool IsEmpty<T>()
             where T : struct, Enum
             => Cache_Values<T>.IsEmpty;
@@ -128,6 +132,7 @@ namespace FastEnumUtility
         /// </summary>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool IsContinuous<T>()
             where T : struct, Enum
             => Cache_UnderlyingOperation<T>.UnderlyingOperation.IsContinuous;
@@ -140,6 +145,7 @@ namespace FastEnumUtility
         /// </summary>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool IsFlags<T>()
             where T : struct, Enum
             => Cache_IsFlags<T>.IsFlags;
@@ -165,6 +171,7 @@ namespace FastEnumUtility
         /// <param name="name"></param>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool IsDefined<T>(string name)
             where T : struct, Enum
             => TryParseName<T>(name, false, out _);
@@ -178,11 +185,12 @@ namespace FastEnumUtility
         /// <param name="value"></param>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns></returns>
-        public static T Parse<T>(string value)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static T Parse<T>(string? value)
             where T : struct, Enum
             => TryParseInternal<T>(value, false, out var result)
             ? result
-            : throw new ArgumentException(nameof(value));
+            : throw new ArgumentException(null, nameof(value));
 
 
         /// <summary>
@@ -192,11 +200,12 @@ namespace FastEnumUtility
         /// <param name="value"></param>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns></returns>
-        public static T Parse<T>(string value, bool ignoreCase)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static T Parse<T>(string? value, bool ignoreCase)
             where T : struct, Enum
             => TryParseInternal<T>(value, ignoreCase, out var result)
             ? result
-            : throw new ArgumentException(nameof(value));
+            : throw new ArgumentException(null, nameof(value));
 
 
         /// <summary>
@@ -207,7 +216,8 @@ namespace FastEnumUtility
         /// <param name="result"></param>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns>true if the value parameter was converted successfully; otherwise, false.</returns>
-        public static bool TryParse<T>(string value, out T result)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool TryParse<T>(string? value, out T result)
             where T : struct, Enum
             => TryParseInternal(value, false, out result);
 
@@ -222,7 +232,8 @@ namespace FastEnumUtility
         /// <param name="result"></param>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns>true if the value parameter was converted successfully; otherwise, false.</returns>
-        public static bool TryParse<T>(string value, bool ignoreCase, out T result)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool TryParse<T>(string? value, bool ignoreCase, out T result)
             where T : struct, Enum
             => TryParseInternal(value, ignoreCase, out result);
 
@@ -237,10 +248,10 @@ namespace FastEnumUtility
         /// <param name="result"></param>
         /// <typeparam name="T">Enum type</typeparam>
         /// <returns></returns>
-        private static bool TryParseInternal<T>(string value, bool ignoreCase, out T result)
+        private static bool TryParseInternal<T>(string? value, bool ignoreCase, out T result)
             where T : struct, Enum
         {
-            if (string.IsNullOrEmpty(value))
+            if (value is null || string.IsNullOrEmpty(value))
             {
                 result = default;
                 return false;
